+++ kamehouse-groot/public/kame-house-groot/js/admin/server-manager/groot-server-manager.js
@@ -219,7 +219,7 @@ class GitManager {
     }
     kameHouse.extension.serverManager.setCommandRunning();
     kameHouse.extension.serverManager.openExecutingCommandModal();
-    kameHouse.extension.kameHouseShell.execute('kamehouse/exec-script-all-servers.sh', "-s common/git/git-pull-kamehouse.sh", false, 600, 
+    kameHouse.extension.kameHouseShell.execute('kamehouse/exec-kamehouse-all-servers.sh', "-s common/git/git-pull-kamehouse.sh", false, 600, 
       (scriptOutput) => kameHouse.extension.serverManager.completeCommandCallback(scriptOutput), 
       (scriptOutput) => kameHouse.extension.serverManager.completeCommandCallback(scriptOutput));
   }
@@ -369,7 +369,7 @@ class DeploymentManager {
     kameHouse.extension.serverManager.setCommandRunning();
     kameHouse.extension.serverManager.openExecutingCommandModal();
     const args = "-s kamehouse/deploy-kamehouse.sh -a -m" + SPACE + module;
-    kameHouse.extension.kameHouseShell.execute('kamehouse/exec-script-all-servers.sh', args, false, 600, (scriptOutput) => this.refreshServerView(scriptOutput), () => {});
+    kameHouse.extension.kameHouseShell.execute('kamehouse/exec-kamehouse-all-servers.sh', args, false, 600, (scriptOutput) => this.refreshServerView(scriptOutput), () => {});
   }
 
   /**
@@ -401,7 +401,7 @@ class DeploymentManager {
     }
     kameHouse.extension.serverManager.setCommandRunning();
     kameHouse.extension.serverManager.openExecutingCommandModal();
-    kameHouse.extension.kameHouseShell.execute('kamehouse/exec-script-all-servers.sh', "-s kamehouse/deploy-kamehouse.sh", false, 600, (scriptOutput) => this.refreshServerView(scriptOutput), () => {});
+    kameHouse.extension.kameHouseShell.execute('kamehouse/exec-kamehouse-all-servers.sh', "-s kamehouse/deploy-kamehouse.sh", false, 600, (scriptOutput) => this.refreshServerView(scriptOutput), () => {});
   }
 
   /**